Fix tests and improve codebase
