+++ react/components/BuyButton/Wrapper.js
@@ -54,10 +54,6 @@ const BuyButtonWrapper = ({
     }
   }
 
-  const { available, ...restProps } = buyButtonProps()
-
-  if (!available) return null
-
   return (
     <BuyButton
       intl={intl}
@@ -66,8 +62,7 @@ const BuyButtonWrapper = ({
       onAddFinish={onAddFinish}
       showToast={showToast}
       orderFormContext={orderFormContext}
-      available={available}
-      { ...restProps }
+      { ...buyButtonProps() }
     >
       <FormattedMessage id="store/buy-button.add-to-cart" />
     </BuyButton>
